refactor(FileListHeader): Group header
