+++ src/modules/views/Public/PublicFolderView.jsx
@@ -274,4 +274,4 @@ const PublicFolderView = () => {
   )
 }
 
-export default PublicFolderView
+export { PublicFolderView }
